requirements.txt
pytesseract
Pillow
gspread
oauth2client
thefuzz
python-dotenv
google-api-python-client
